// src/app/modules/information/models/job-description.ts
export class JobDescription{
    _id?: string;
    functions?: string;
    contractType?: string;
    duration?: string;
    location?: string;
    workDayType?: string;
    schedule?: string;
    salary?: number;
    requiredTitle?: string;
    computerRequirements?: string;
    languagesRequirements?: string;
    driverLicense?: boolean;
    carOwner?: boolean;
    age?: number;
    incorporationDate?: Date;
}